% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{get_data}
\usage{
get_file(node, remote_path = NULL, file, path = ".", conflicts = "overwrite")
}
\arguments{
\item{node}{character, osf node}

\item{remote_path}{character, remote path to file}

\item{file}{character, name of file to download}

\item{path}{character, output path. Defaults to working directory}

\item{conflicts}{What to do if a file already exists. Default is "overwrite" see \code{\link[osfr]{osf_download}} for other options.}
}
\description{
get_data
}
\details{
logic
!file exists - download
file exists & !matching hashes - download
}
\examples{
get_file(node = "p7ayb", file = "Analysis notes.txt", path = "data")
}
