% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_update.R
\name{need_update}
\alias{need_update}
\title{Need Update}
\usage{
need_update(node, remote_path = NULL, file, path = ".")
}
\arguments{
\item{node}{character, osf node}

\item{remote_path}{character, remote path to file}

\item{file}{character, name of file to download}

\item{path}{character, output path. Defaults to working directory}
}
\description{
Need Update
}
\details{
logic
!file exists - download
file exists & !matching hashes - download
}
\examples{
need_update(node = "p7ayb", file = "Analysis notes.txt", path = "data")
}
