% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_file}
\alias{get_file}
\title{get_data}
\usage{
get_file(file, path = ".")
}
\arguments{
\item{file}{character, name of file to download}

\item{path}{character, output path. Defaults to working directory}
}
\description{
get_data
}
\details{
logic
}
