% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{get_data}
\usage{
get_file(node, file, path = ".")
}
\arguments{
\item{node}{character, osf node}

\item{file}{character, name of file to download}

\item{path}{character, output path. Defaults to working directory}
}
\description{
get_data
}
\details{
logic
!file exists - download
file exists & !matching hashes - download
}
\examples{
get_file(node = "p7ayb", file = "Analysis notes.txt", path = "data")
}
